Modify the text widget sample to simulate column spanning by using an embedded panel. This is temoprary until I implement column and row spanning. Fix gridbaglayout.cpp to compile with gcc / mingw. Removed "converting float to int" warnings.

+++ trunk/include/cpaf/gui/gridbaglayout.h
@@ -253,7 +253,7 @@ private:
         Calculates the sizes of the widgets contained in this objects columns or rows
         based on the template parameter
     */
-    template<gblm::GROUP> void calc_group_sizes(int avail, WidgetRects &rects);
+    template<gblm::GROUP> void calc_group_sizes(float avail, WidgetRects &rects);
 
     /*!
         \return Weights for the widgets in the given row or column based on the
